Добавила заготовки для страниц
