+++ package.json
@@ -19,6 +19,8 @@
     "react-dnd-html5-backend": "^16.0.1",
     "react-dom": "^18.3.1",
     "react-redux": "^9.1.2",
+    "react-router": "^6.27.0",
+    "react-router-dom": "^6.27.0",
     "react-scripts": "5.0.1",
     "redux": "^5.0.1",
     "redux-thunk": "^3.1.0",
