+++ src/pages/forgot-password/forgot-password.tsx
@@ -1,0 +1,3 @@
+export function ForgotPasswordPage() {
+  return <></>;
+}
