+++ src/pages/ingredients/ingredients.tsx
@@ -1,0 +1,3 @@
+export function IngredientsPage() {
+  return <></>;
+}
