+++ src/pages/profile/profile.tsx
@@ -1,0 +1,3 @@
+export function ProfilePage() {
+  return <></>;
+}
