+++ src/pages/register/register.tsx
@@ -1,0 +1,3 @@
+export function RegisterPage() {
+  return <></>;
+}
