+++ src/pages/reset-password/reset-password.tsx
@@ -1,0 +1,3 @@
+export function ResetPasswordPage() {
+  return <></>;
+}
